refactor: Rename `useIsConnected` to `useNetworkConnectivity`
Attempt to better communicate the Hook intent.
