+++ test/native/setup.js
@@ -107,7 +107,8 @@ jest.mock( '@wordpress/react-native-bridge', () => {
 		subscribeShowEditorHelp: jest.fn(),
 		subscribeOnUndoPressed: jest.fn(),
 		subscribeOnRedoPressed: jest.fn(),
-		useIsConnected: jest.fn( () => ( { isConnected: true } ) ),
+		subscribeConnectionStatus: jest.fn( () => ( { remove: jest.fn() } ) ),
+		requestConnectionStatus: jest.fn(),
 		editorDidMount: jest.fn(),
 		editorDidAutosave: jest.fn(),
 		subscribeMediaUpload: jest.fn(),
